adjusting the remapping mechanism to cache, and avoid leaking blue globals that are not DOM APIs
