+++ src/environment.ts
@@ -1,25 +1,18 @@
 import { 
-    apply, 
+    apply,
     assign,
-    isUndefined, 
-    isNullOrUndefined,
-    ObjectCreate, 
-    isFunction, 
-    hasOwnProperty, 
-    ReflectDefineProperty, 
+    isUndefined,
+    ObjectCreate,
+    isFunction,
+    ReflectDefineProperty,
     emptyArray,
     ErrorCreate,
-    ESGlobalKeys,
-    ReflectGetOwnPropertyDescriptor, 
-    ReflectIsExtensible,
-    SetHas,
+    ReflectGetOwnPropertyDescriptor,
     WeakMapCreate,
-    WeakMapHas,
     WeakMapSet,
-    ReflectiveIntrinsicObjectNames,
     WeakMapGet,
     construct,
-    ownKeys,
+    isTrue,
 } from './shared';
 import { serializedRedEnvSourceText, MarshalHooks } from './red';
 import { blueProxyFactory } from './blue';
@@ -48,29 +41,6 @@ interface SecureEnvironmentOptions {
     distortionMap?: Map<RedProxyTarget, RedProxyTarget>;
 }
 
-// TODO: type this better based on ReflectiveIntrinsicObjectNames
-type ReflectiveIntrinsicsMap = Record<string, any>;
-
-const cachedReflectiveIntrinsicsMap: WeakMap<typeof globalThis, ReflectiveIntrinsicsMap> = WeakMapCreate();
-
-function getReflectiveIntrinsics(global: typeof globalThis): ReflectiveIntrinsicsMap {
-    let reflectiveIntrinsics: ReflectiveIntrinsicsMap | undefined = WeakMapGet(cachedReflectiveIntrinsicsMap, global);
-    if (!isUndefined(reflectiveIntrinsics)) {
-        return reflectiveIntrinsics;
-    }
-    reflectiveIntrinsics = ObjectCreate(null) as ReflectiveIntrinsicsMap;
-    WeakMapSet(cachedReflectiveIntrinsicsMap, global, reflectiveIntrinsics);
-    // remapping intrinsics that are realm's agnostic
-    for (let i = 0, len = ReflectiveIntrinsicObjectNames.length; i < len; i += 1) {
-        const name = ReflectiveIntrinsicObjectNames[i];
-        reflectiveIntrinsics[name] = global[name];
-    }
-    return reflectiveIntrinsics;
-}
-
-// caching from the blue realm right away to avoid picking up modified entries
-getReflectiveIntrinsics(globalThis);
-
 export class SecureEnvironment implements MembraneBroker {
     // map from red to blue references
     redMap: WeakMap<RedFunction | RedObject, RedProxyTarget | BlueProxy> = WeakMapCreate();
@@ -83,7 +53,7 @@ export class SecureEnvironment implements MembraneBroker {
         if (isUndefined(options)) {
             throw ErrorCreate(`Missing SecureEnvironmentOptions options bag.`);
         }
-        const { blueGlobalThis, redGlobalThis, distortionMap } = options;
+        const { redGlobalThis, distortionMap } = options;
         this.distortionMap = WeakMapCreate();
         // validating distortion entries
         distortionMap?.forEach((value, key) => {
@@ -105,16 +75,6 @@ export class SecureEnvironment implements MembraneBroker {
         };
         this.getRedValue = redEnvFactory(this, blueHooks);
         this.getBlueValue = blueProxyFactory(this);
-        // remapping intrinsics that are realm's agnostic
-        const blueIntrinsics = getReflectiveIntrinsics(blueGlobalThis);
-        const redIntrinsics = getReflectiveIntrinsics(redGlobalThis);
-        for (let i = 0, len = ReflectiveIntrinsicObjectNames.length; i < len; i += 1) {
-            const name = ReflectiveIntrinsicObjectNames[i];
-            const blue = blueIntrinsics[name];
-            const red = redIntrinsics[name];
-            this.setRefMapEntries(red, blue);
-            this.setRefMapEntries(red.prototype, blue.prototype);
-        }
     }
 
     getBlueValue(red: RedValue): BlueValue {
@@ -145,117 +105,53 @@ export class SecureEnvironment implements MembraneBroker {
         WeakMapSet(this.blueMap, blue, red);
     }
 
-    remap(redValue: RedValue, blueValue: BlueValue, blueDescriptors?: PropertyDescriptorMap) {
-        this.setRefMapEntries(redValue, blueValue);
-        if (!isUndefined(blueDescriptors)) {
-            const keys = ownKeys(blueDescriptors);
-            for (let i = 0, len = keys.length; i < len; i += 1) {
-                const key = keys[i];
-                // TODO: this whole loop needs cleanup and simplification avoid
-                // overriding ECMAScript global keys.
-                if (SetHas(ESGlobalKeys, key) || !hasOwnProperty(blueDescriptors, key)) {
-                    continue;
-                }
+    remap(redValue: RedValue, blueValue: BlueValue, blueDescriptors: PropertyDescriptorMap) {
+        const broker = this;
+        for (const key in blueDescriptors) {
+            if (!canRedPropertyBeTamed(redValue, key)) {
+                console.warn(`Property ${key} of ${redValue} cannot be remapped.`);
+                continue;
+            }
+            // avoid poisoning by only installing own properties from blueDescriptors
+            const blueDescriptor = assign(ObjectCreate(null), blueDescriptors[key]);
+            const redDescriptor = assign(ObjectCreate(null), blueDescriptor);
+            if ('value' in blueDescriptor) {
+                redDescriptor.value = broker.getRedValue(blueDescriptor.value);
+            } else {
+                // Use the original getter to return a red object, but if the
+                // sandbox attempts to set it to a new value, this mutation will
+                // only affect the sandbox's global object, and the getter will
+                // start returning the new provided value rather than calling onto
+                // the blue realm. This is to preserve the object graph of the
+                // blue realm.
+                let currentBlueGetter: (this: RedValue) => RedValue = () => undefined;
 
-                // avoid poisoning by only installing own properties from blueDescriptors
-                // @ts-ignore PropertyDescriptorMap def defines properties as being only of string type
-                const blueDescriptor = assign(ObjectCreate(null), blueDescriptors[key]);
-                if ('value' in blueDescriptor) {
-                    // TODO: maybe we should make everything a getter/setter that way
-                    // we don't pay the cost of creating the proxy in the first place
-                    blueDescriptor.value = this.getRedValue(blueDescriptor.value);
-                } else {
-                    // Use the original getter to return a red object, but if the 
-                    // sandbox attempts to set it to a new value, this mutation will
-                    // only affect the sandbox's global object, and the getter will
-                    // start returning the new provided value rather than calling onto
-                    // the blue realm. This is to preserve the object graph of the
-                    // blue realm.
-                    const env = this;
-                    const { get: originalGetter } = blueDescriptor;
-
-                    let currentGetter = () => undefined;
-                    if (isFunction(originalGetter)) {
-                        const originalOrDistortedGetter: () => any = WeakMapGet(this.distortionMap, originalGetter) || originalGetter;
-                        currentGetter = function(this: any): RedValue {
-                            const value: BlueValue = apply(originalOrDistortedGetter, env.getBlueValue(this), emptyArray);
-                            return env.getRedValue(value);
-                        };
-                    }
-
-                    blueDescriptor.get = function(): RedValue {
-                        return apply(currentGetter, this, emptyArray);
+                if (isFunction(blueDescriptor.get)) {
+                    const { get: blueGetter } = blueDescriptor;
+                    const blueDistortedGetter: () => BlueValue = WeakMapGet(this.distortionMap, blueGetter) || blueGetter;
+                    currentBlueGetter = function() {
+                        const value: BlueValue = apply(blueDistortedGetter, broker.getBlueValue(this), emptyArray);
+                        return broker.getRedValue(value);
                     };
-
-                    if (isFunction(blueDescriptor.set)) {
-                        blueDescriptor.set = function(v: RedValue): void {
-                            // if a global setter is invoke, the value will be use as it is as the result of the getter operation
-                            currentGetter = () => v;
-                        };
-                    }
+                    redDescriptor.get = function(): RedValue {
+                        return apply(currentBlueGetter, this, emptyArray);
+                    };
                 }
 
-                const redDescriptor = ReflectGetOwnPropertyDescriptor(redValue, key);
-                if (!isUndefined(redDescriptor) && 
-                        hasOwnProperty(redDescriptor, 'configurable') &&  
-                        redDescriptor.configurable === false) {
-                    const redPropertyValue = redValue[key];
-                    if (isNullOrUndefined(redPropertyValue)) {
-                        continue;
-                    }
-                    // this is the case where the red realm has a global descriptor that was supposed to be
-                    // overrule but can't be done because it is a non-configurable. Instead we try to
-                    // fallback to some more advanced gymnastics
-                    if (hasOwnProperty(redDescriptor, 'value')) {
-                        // valid proxy target (intentionally ignoring the case of document.all since it is not a value descriptor)
-                        if (typeof redPropertyValue === 'function' || typeof redPropertyValue === 'object') {
-                            if (!WeakMapHas(this.redMap, redPropertyValue)) {
-                                // remapping the value of the red object graph to the blue realm graph
-                                const { value: blueDescriptorValue } = blueDescriptor;
-                                if (redValue !== blueDescriptorValue) {
-                                    if (this.getBlueValue(redValue) !== blueValue) {
-                                        console.error('need remapping: ',  key, blueValue, blueDescriptor);
-                                    } else {
-                                        // it was already mapped
-                                    }
-                                } else {
-                                    // window.top is the classic example of a descriptor that leaks access to the blue
-                                    // window reference, and there is no containment for that case yet.
-                                    console.error('leaking: ',  key, blueValue, blueDescriptor);
-                                }
-                            } else {
-                                // an example of this is circular window.window ref
-                                console.info('circular: ',  key, blueValue, blueDescriptor);
-                            }
-                        }
-                    } else if (hasOwnProperty(redDescriptor, 'get')) {
-                        // internationally ignoring the case of (typeof document.all === 'undefined') because
-                        // it is specified as configurable, you never get one of those exotic objects in this branch
-                        if (typeof redPropertyValue === 'function' || typeof redPropertyValue === 'object') {
-                            if (redPropertyValue === redValue[key]) {
-                                // this is the case for window.document which is identity preserving getter
-                                // const blueDescriptorValue = blueValue[key];
-                                // this.setRefMapEntries(redDescriptorValue, blueDescriptorValue);
-                                // this.installDescriptors(redDescriptorValue, blueDescriptorValue, getOwnPropertyDescriptors(blueDescriptorValue));
-                                console.error('need remapping: ', key, blueValue, blueDescriptor);
-                                if (ReflectIsExtensible(redPropertyValue)) {
-                                    // remapping proto chain
-                                    // ReflectSetPrototypeOf(redDescriptorValue, this.getRedValue(ReflectGetPrototypeOf(redDescriptorValue)));
-                                    console.error('needs prototype remapping: ', key, blueValue);
-                                } else {
-                                    console.error('leaking prototype: ',  key, blueValue, blueDescriptor);
-                                }
-                            } else {
-                                console.error('leaking a getter returning values without identity: ', key, blueValue, blueDescriptor);
-                            }
-                        } else {
-                            console.error('skipping: ', key, blueValue, blueDescriptor);
-                        }
-                    }
-                } else {
-                    ReflectDefineProperty(redValue, key, blueDescriptor);
+                if (isFunction(blueDescriptor.set)) {
+                    redDescriptor.set = function(v: RedValue): void {
+                        // if a global setter is invoke, the value will be use as it is as the result of the getter operation
+                        currentBlueGetter = () => v;
+                    };
                 }
             }
+            ReflectDefineProperty(redValue, key, redDescriptor);
         }
     }
+}
+
+function canRedPropertyBeTamed(redValue: RedValue, key: PropertyKey): boolean {
+    const redDescriptor = ReflectGetOwnPropertyDescriptor(redValue, key);
+    // TODO: what about writable - non-configurable?
+    return isUndefined(redDescriptor) || isTrue(redDescriptor.configurable);
 }
