+++ src/node-realm.ts
@@ -1,7 +1,7 @@
 import { SecureEnvironment } from "./environment";
 import { RedProxyTarget } from "./types";
-import { getOwnPropertyDescriptors } from "./shared";
 import { runInNewContext } from 'vm';
+import { getFilteredEndowmentDescriptors, linkIntrinsics } from "./intrinsics";
 
 // note: in a node module, the top-level 'this' is not the global object
 // (it's *something* but we aren't sure what), however an indirect eval of
@@ -11,6 +11,7 @@ const unsafeGlobalEvalSrc = `(0, eval)("'use strict'; this")`;
 export default function createSecureEnvironment(distortionMap?: Map<RedProxyTarget, RedProxyTarget>, endowments?: object): (sourceText: string) => void {
     // Use unsafeGlobalEvalSrc to ensure we get the right 'this'.
     const redGlobalThis = runInNewContext(unsafeGlobalEvalSrc);
+    const endowmentsDescriptors = getFilteredEndowmentDescriptors(endowments || {});
     const { eval: redIndirectEval } = redGlobalThis;
     const blueGlobalThis = globalThis as any;
     const env = new SecureEnvironment({
@@ -18,9 +19,10 @@ export default function createSecureEnvironment(distortionMap?: Map<RedProxyTarg
         redGlobalThis,
         distortionMap,
     });
+    linkIntrinsics(env, blueGlobalThis, redGlobalThis);
 
     // remapping globals
-    env.remap(redGlobalThis, blueGlobalThis, endowments && getOwnPropertyDescriptors(endowments));
+    env.remap(redGlobalThis, blueGlobalThis, endowmentsDescriptors);
 
     return (sourceText: string): void => {
         try {
