+++ src/shared.ts
@@ -80,88 +80,3 @@ export function isFunction(obj: any): obj is Function {
 }
 
 export const emptyArray: [] = [];
-
-export const ESGlobalKeys = SetCreate([
-
-    // *** 18.1 Value Properties of the Global Object
-    'Infinity',
-    'NaN',
-    'undefined',
-
-    // *** 18.2 Function Properties of the Global Object
-    'eval', // dangerous
-    'isFinite',
-    'isNaN',
-    'parseFloat',
-    'parseInt',
-    'decodeURI',
-    'decodeURIComponent',
-    'encodeURI',
-    'encodeURIComponent',
-
-    // *** 18.3 Constructor Properties of the Global Object
-    'Array',
-    'ArrayBuffer',
-    'Boolean',
-    'DataView',
-    'Date', // Unstable
-    'Error', // Unstable
-    'EvalError',
-    'Float32Array',
-    'Float64Array',
-    'Function', // dangerous
-    'Int8Array',
-    'Int16Array',
-    'Int32Array',
-    'Map',
-    'Number',
-    'Object',
-    // Allow Blue `Promise` constructor to overwrite the Red one so that promises
-    // created by the `Promise` constructor or APIs like `fetch` will work.
-    //'Promise',
-    'Proxy', // Unstable
-    'RangeError',
-    'ReferenceError',
-    'RegExp', // Unstable
-    'Set',
-    'SharedArrayBuffer',
-    'String',
-    'Symbol',
-    'SyntaxError',
-    'TypeError',
-    'Uint8Array',
-    'Uint8ClampedArray',
-    'Uint16Array',
-    'Uint32Array',
-    'URIError',
-    'WeakMap',
-    'WeakSet',
-
-    // *** 18.4 Other Properties of the Global Object
-    'Atomics',
-    'JSON',
-    'Math',
-    'Reflect',
-
-    // *** Annex B
-    'escape',
-    'unescape',
-
-    // *** ECMA-402
-    'Intl', // Unstable
-]);
-
-// These are foundational things that should never be wrapped but are equivalent
-// TODO: revisit this list.
-export const ReflectiveIntrinsicObjectNames = [
-    'Array',
-    'Object',
-    'Function',
-    'URIError',
-    'TypeError',
-    'SyntaxError',
-    'ReferenceError',
-    'RangeError',
-    'EvalError',
-    'Error',
-];
